using long int for memory for now
git-svn-id: https://sirius.chem.vt.edu/svn/psi4/trunk@252 6a214465-e6f0-41bc-91e6-de896bb96b5e

Former-commit-id: 34c4a0df8e6a077d1dbd8ea269389d4d2fac9654

+++ include/psi4-dec.h
@@ -11,14 +11,14 @@ namespace psi {
 
   class Module {
     std::string prgid;
-    size_t memory;
+    long int memory;
 
   public:
     Module(std::string s = "PSI4") { prgid = s; }
     void set_prgid (std::string s) { prgid = s; }
     std::string gprgid(void) const { return prgid; }
-    size_t get_memory(void) const { return memory; }
-    void set_memory(size_t newmem) { memory = newmem; }
+    long int get_memory(void) const { return memory; }
+    void set_memory(long int newmem) { memory = newmem; }
   };
 
   struct ModuleInformation {
